страницы регистрации и авторизации - единая точка входа, единый скрипт
