+++ src/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+<head id="head">
+    <meta charset="UTF-8">
+    <title>Мессенджер</title>
+</head>
+<body>
+<div id='root'></div>
+</body>
+<script type="module" src="app.ts"></script>
+</html>
