--- src/pages/authorization/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Авторизация</title>
-    <link rel='stylesheet' href='./authorization.styles.scss'>
-</head>
-<body>
-<div id='root'></div>
-</body>
-<script type="module" src="./script.ts"></script>
-</html>
