+++ src/pages/registry/registry.helpers.ts
@@ -1,5 +1,7 @@
 import {triggerValidateError} from "../../utils/FormValidation";
 import {FIO_MASK, LOGIN_MASK, PASSWORD_REGEXP, PHONE_MASK} from "../../utils/Masks";
+import Router from "../../modules/router";
+import {addFocusEventOnInput} from "../../utils/FormEvents";
 
 export const addFormSubmitEvent = (registryForm: HTMLFormElement) => {
     registryForm.onsubmit = (event) => {
@@ -140,4 +142,27 @@ const validate = (registryForm: HTMLFormElement) => {
     }
 
     return result;
+}
+
+export const prepareRegistrationForm = (registryForm: HTMLFormElement, router: Router) => {
+
+    // события на фокус
+    addFocusEventOnInput(registryForm?.login, ['login-empty', 'login-error']);
+    addFocusEventOnInput(registryForm?.mail, ['mail-empty']);
+    addFocusEventOnInput(registryForm?.firstName, ['first-name-empty', 'first-name-error']);
+    addFocusEventOnInput(registryForm?.secondName, ['second-name-empty', 'second-name-error']);
+    addFocusEventOnInput(registryForm?.phone, ['phone-empty', 'phone-error']);
+    addFocusEventOnInput(registryForm?.password, ['password-empty', 'password-error']);
+    addFocusEventOnInput(registryForm?.passwordRepeat,
+        ['password-repeat-empty', 'password-repeat-error', 'password-repeat-equality']);
+
+    if(registryForm) {
+        addFormSubmitEvent(registryForm);
+    }
+
+    const authHref = document.querySelector('#enterHref');
+    authHref?.addEventListener('click', (e) => {
+        e.preventDefault();
+        router.go('/')
+    })
 }
